fix: banded ladder promoted app-wide - money tab swapped, old component retired
