+++ frontend/src/components/dashboard/RateTargetsCard.tsx
@@ -1,5 +1,5 @@
 import type { useRateTargets } from "@/hooks/useRateTargets";
-import { RateLadder } from "./RateLadder";
+import { BandedLadder } from "@/components/ui/BandedLadder";
 import { Panel } from "@/components/ui/Panel";
 import { money } from "@/lib/format";
 
@@ -163,7 +163,8 @@ export const RateTargetsCard = ({ targets }: { targets: Targets }) => {
             <p className="text-xs text-muted-text mb-3">
               Rate to book · gross $/mile driven · marker = your rate
             </p>
-            <RateLadder ladder={bookingLadder} rpm={grossRate} spec={specLadder} />
+            <BandedLadder ladder={bookingLadder} rpm={grossRate} label="Standard flatbed" />
+            <BandedLadder ladder={specLadder} rpm={grossRate} label="Oversize / specialized" />
             <p className="text-[11px] text-muted-text mt-2">
               walk-away = your cost/mile ÷ your{" "}
               {Math.round(linehaulTake * 100)}% keep. Book above it — with your
